fix: mishandling of version and packageversion (#115)
This PR addresses multiple issues with the newly added `PackageVersion`
field, as reported in #113 and #114.

- The distinction between using `version` and `packageversion` was
missing from the `VersionCli.cs` which is responsible for some actions.
The usage of the new field has been fixed
- The `ProjectFileParser` was likewise missing proper handling of the
new field and thus defaulted to using version / versionprefix
- The `dump version` command was blatantly hardcoded to just dumping
`PackageVersion` which is obviously wrong. It should now respect the CLI
settings of using either `Version` or `PackageVersion` for dumping.